pyproject.toml
[project]
name = "hushh-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno[google]>=1.4.0",
    "fastapi>=0.115.0",
    "google-genai>=1.65.0",
    "pgvector>=0.3.0",
    "psycopg[binary]>=3.2.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.30.0",
]
